feat(private-rpc): moved files
refactor. Moved classes around to have a more cohesive file structure.
