+++ packages/private-rpc/src/build-app.ts
@@ -7,7 +7,8 @@ import {
 import { DB } from '@/db';
 import { usersRoutes } from '@/routes/users-routes';
 import { rpcRoutes } from '@/routes/rpc-routes';
-import { Authorizer } from '@/permissions/group';
+
+import { Authorizer } from '@/permissions/authorizeer';
 
 export function buildApp(
   produceLogs = true,
